+++ src/components/Options.jsx
@@ -1,15 +1,7 @@
-export default function Options() {
+export default function Options({ color }) {
   return (
     <>
-      <option value="">-- Choose --</option>
-      <option value="red">Red</option>
-      <option value="dark">Dark</option>
-      <option value="orange">Orange</option>
-      <option value="yellow">Yellow</option>
-      <option value="green">Green</option>
-      <option value="blue">Blue</option>
-      <option value="indigo">Indigo</option>
-      <option value="violet">Violet</option>
+      <option value={color}>{color}</option>
     </>
   );
 }